// core/lib/Lavenza/Client/ClientType.ts
/**
 * Project Lavenza
 * Copyright 2017-2019 Aigachu, All Rights Reserved
 *
 * License: https://github.com/Aigachu/Lavenza-II/blob/master/LICENSE
 */

/**
 * A little Enum to manage our types of Clients.
 *
 * We don't wanna have to change them EVERYWHERE. So we manage them all here!
 */
enum ClientType {
  Discord = "discord",
  Twitch = "twitch",
  // Slack = 'slack'
  // Youtube = 'youtube',
  // Skype = 'a funny joke',
  // Whatsapp = 'oh really?',
  // Messenger = 'LOL! This one will be fun...',
  // WeChat = 'CHINESE TRANSLATIONS?!',
}

export { ClientType };